+++ lua/custom/plugins/guess-indent.lua
@@ -1,0 +1,4 @@
+-- Detect tabstop and shiftwidth automatically
+return {
+  'NMAC427/guess-indent.nvim',
+}
