+++ lua/custom/plugins/todo-comments.lua
@@ -1,0 +1,7 @@
+-- Highlight todo, notes, etc in comments
+return {
+  'folke/todo-comments.nvim',
+  event = 'VimEnter',
+  dependencies = { 'nvim-lua/plenary.nvim' },
+  opts = { signs = false },
+}
