+++ lua/custom/plugins/tresitter-context.lua
@@ -1,0 +1,5 @@
+return {
+  'nvim-treesitter/nvim-treesitter-context',
+  dependencies = { 'nvim-treesitter/nvim-treesitter' },
+  event = { 'BufReadPost', 'BufNewFile' },
+}
